[Front] ESG 대시보드 연동 및 웹페이지 크기에 따른 크기 수정
