+++ src/components/weather/WeatherRiskPanel.vue
@@ -15,8 +15,8 @@ defineProps({
 </script>
 
 <template>
-  <div class="flex h-[744px] flex-col overflow-hidden rounded-2xl border border-forena-100/90 bg-white/95 shadow-card">
-    <div class="border-b border-forena-100 bg-gradient-to-r from-violet-50/70 via-white to-rose-50/60 px-5 py-4">
+  <div class="flex h-[588px] flex-col overflow-hidden rounded-2xl border border-forena-100/90 bg-white/95 shadow-card">
+    <div class="border-b border-forena-100 bg-gradient-to-r from-violet-50/70 via-white to-rose-50/60 px-5 py-3">
       <div class="flex flex-wrap items-start justify-between gap-2">
         <div>
           <h2 class="flex items-center gap-2 text-base font-extrabold text-forena-900">
@@ -32,12 +32,12 @@ defineProps({
       </div>
     </div>
 
-    <div class="flex flex-1 flex-col gap-4 overflow-hidden p-4 md:p-5">
+    <div class="flex flex-1 flex-col gap-3 overflow-hidden p-4">
       <div
-        class="relative flex min-h-0 flex-1 flex-col overflow-hidden rounded-2xl border-2 border-rose-200/80 bg-gradient-to-br from-rose-50/80 via-white to-rose-50/50 p-5 shadow-sm"
+        class="relative flex min-h-0 flex-1 flex-col overflow-hidden rounded-2xl border-2 border-rose-200/80 bg-gradient-to-br from-rose-50/80 via-white to-rose-50/50 p-4 shadow-sm"
       >
         <span class="pointer-events-none absolute left-0 top-0 h-full w-1.5 bg-gradient-to-b from-rose-400 to-rose-600" />
-        <div class="mb-4 flex items-center gap-3">
+        <div class="mb-3 flex items-center gap-3">
           <span class="rounded-lg bg-rose-100 px-3.5 py-2 text-sm font-extrabold text-rose-800">AI</span>
           <AlertTriangle class="h-5 w-5 text-rose-600" />
           <h3 class="text-[18px] font-extrabold tracking-tight text-forena-900">{{ WEATHER_CONTROL_TEXT.row3Title }}</h3>
@@ -46,11 +46,11 @@ defineProps({
           </span>
         </div>
 
-        <div v-if="planRisks.length > 0" class="min-h-0 flex-1 space-y-3 overflow-y-auto pr-1">
+        <div v-if="planRisks.length > 0" class="min-h-0 flex-1 space-y-2.5 overflow-y-auto pr-1">
           <div
             v-for="risk in planRisks"
             :key="`plan-${risk.title}-${risk.reason}`"
-            class="rounded-xl border border-rose-100 bg-white/70 px-4 py-3"
+            class="rounded-xl border border-rose-100 bg-white/70 px-4 py-2.5"
           >
             <div class="mb-1 flex items-center justify-between gap-2">
               <p class="text-[14px] font-extrabold leading-5 text-forena-900">{{ risk.title }}</p>
@@ -71,10 +71,10 @@ defineProps({
       </div>
 
       <div
-        class="relative flex min-h-0 flex-1 flex-col overflow-hidden rounded-2xl border-2 border-violet-200/80 bg-gradient-to-br from-violet-50/80 via-white to-violet-50/40 p-5 shadow-sm"
+        class="relative flex min-h-0 flex-1 flex-col overflow-hidden rounded-2xl border-2 border-violet-200/80 bg-gradient-to-br from-violet-50/80 via-white to-violet-50/40 p-4 shadow-sm"
       >
         <span class="pointer-events-none absolute left-0 top-0 h-full w-1.5 bg-gradient-to-b from-violet-400 to-violet-600" />
-        <div class="mb-4 flex items-center gap-3">
+        <div class="mb-3 flex items-center gap-3">
           <span class="rounded-lg bg-violet-100 px-3.5 py-2 text-sm font-extrabold text-violet-800">AI</span>
           <Sparkles class="h-5 w-5 text-violet-600" />
           <h3 class="text-[18px] font-extrabold tracking-tight text-forena-900">{{ WEATHER_CONTROL_TEXT.row2Title }}</h3>
@@ -83,11 +83,11 @@ defineProps({
           </span>
         </div>
 
-        <div v-if="equipmentRisks.length > 0" class="min-h-0 flex-1 space-y-3 overflow-y-auto pr-1">
+        <div v-if="equipmentRisks.length > 0" class="min-h-0 flex-1 space-y-2.5 overflow-y-auto pr-1">
           <div
             v-for="risk in equipmentRisks"
             :key="`equipment-${risk.title}-${risk.reason}`"
-            class="rounded-xl border border-violet-100 bg-white/70 px-4 py-3"
+            class="rounded-xl border border-violet-100 bg-white/70 px-4 py-2.5"
           >
             <div class="mb-1 flex items-center justify-between gap-2">
               <p class="text-[14px] font-extrabold leading-5 text-forena-900">{{ risk.title }}</p>
